+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [upcoming]
+
+### Changed
+
+- Improve "Skipping CHANNEL_NAME, no new videos to download" to be precise.
+
+### Fixed
+
+- Regex issue with video title. It worked anyway, but better to fix it.
+
 ## [1.1.4] - 2020-11-20
 
 ### Fixed